#141 Enable clipboard actions on the browser's Edit menu
